# frontend/src/api/generated/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "CapabilityProfileListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "CapabilityProfileVersion"
            },
            "type": "array"
          },
          "success": {
            "const": true,
            "type": "boolean"
          },
          "total": {
            "type": "integer"
          }
        },
        "required": [
          "success",
          "items"
        ],
        "type": "object"
      },
      "CapabilityProfileRequest": {
        "properties": {
          "capabilities": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array"
          },
          "profile_id": {
            "type": "string"
          },
          "verified": {
            "type": "boolean"
          },
          "version": {
            "type": "string"
          }
        },
        "required": [
          "profile_id",
          "version"
        ],
        "type": "object"
      },
      "CapabilityProfileResponse": {
        "properties": {
          "item": {
            "$ref": "CapabilityProfileVersion"
          },
          "success": {
            "const": true,
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "type": "object"
      },
...
      "CreateIntentRequest": {
        "properties": {
          "derived_from": {
            "type": "string"
          },
          "profile": {
            "type": "string"
          },
          "project": {
            "type": "string"
          },
          "prompt": {
            "type": "string"
          },
          "ref_slots": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array"
          },
          "seed": {
            "type": "integer"
          },
          "workflow_version": {
            "type": "string"
          }
        },
        "required": [
          "project",
          "prompt"
        ],
        "type": "object"
      },
...
      "DecisionListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "DecisionResponse"
            },
            "type": "array"
          },
          "success": {
            "const": true,
            "type": "boolean"
          },
          "total": {
            "type": "integer"
          }
        },
        "required": [
          "success",
          "items"
        ],
        "type": "object"
      },
...
      "DecisionResponse": {
        "properties": {
          "approval": {
            "$ref": "DecisionRecord"
          },
          "media_verified": {
            "description": "是否真的按磁盘现状核验过。false = 无法验证，**不得**当成「没变」。",
            "type": "boolean"
          },
          "selection": {
            "$ref": "DecisionRecord"
          },
          "state": {
            "description": "采用/批准五态。铁律：selected 永不显示为 approved（ADR-0002）",
            "enum": [
              "selected",
              "selected_not_approved",
              "approved",
              "approved_stale",
              "approved_unverified",
              "none"
            ],
            "type": "string"
          },
          "success": {
            "const": true,
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "state"
        ],
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "description": "稳定错误码，代码里以此分支，不要解析文案",
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "success": {
            "const": false,
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "error",
          "code"
        ],
        "type": "object"
      },
...
      "IntentListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "GenerationIntent"
            },
            "type": "array"
          },
          "success": {
            "const": true,
            "type": "boolean"
          },
          "total": {
            "type": "integer"
          }
        },
        "required": [
          "success",
          "items"
        ],
        "type": "object"
      },
...
      "RevokeApprovalRequest": {
        "properties": {
          "reason": {
            "type": "string"
          },
          "revoked_by": {
            "type": "string"
          }
        },
        "required": [
          "revoked_by",
          "reason"
        ],
        "type": "object"
      },
...
      "SelectMediaRequest": {
        "description": "采用这版。创作决定 —— **不会**自动产生批准。",
        "properties": {
          "note": {
            "type": "string"
          },
          "selected_by": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "本规格是**前端与后端的唯一契约**。前端客户端由 `scripts/generate_client.py` 从本文件生成，`scripts/check_generated_client.py` 在 CI 校验生成物是否过期。\n\n**手工改这个文件没有意义**：它是生成物，改了会被下一次生成覆盖。",
    "title": "漫剧锻造台 API",
    "version": "2026-10-07.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/production_facts/approvals/{approval_id}/revoke": {
      "post": {
        "operationId": "revokeProductionApproval",
        "parameters": [
          {
            "description": "批准记录 id",
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            },
            "description": "成功"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "撤销者非人工主体"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "批准记录不存在"
          }
        },
        "summary": "撤销人工批准（须人工授权者）",
        "tags": [
          "production_facts"
        ],
        "x-contract-source": "manual"
      }
    },
    "/api/production_facts/capability-profiles": {
      "get": {
        "operationId": "listCapabilityProfiles",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilityProfileListResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "能力档案版本列表",
        "tags": [
          "production_facts"
        ],
        "x-contract-source": "manual"
      },
      "post": {
        "operationId": "registerCapabilityProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CapabilityProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilityProfileResponse"
                }
              }
            },
            "description": "成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "参数非法"
          }
        },
        "summary": "登记能力档案版本",
        "tags": [
          "production_facts"
        ],
        "x-contract-source": "manual"
      }
    },
    "/api/production_facts/decisions": {
      "get": {
        "operationId": "listDecisions",
        "parameters": [
          {
            "description": "按媒体版本过滤",
            "in": "query",
            "name": "media_version_id",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionListResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "采用/批准决策列表（decision_state 五态）",
        "tags": [
          "production_facts"
        ],
        "x-contract-source": "manual"
      }
    },
    "/api/production_facts/decisions/history": {
      "get": {
        "operationId": "listDecisionHistory",
        "parameters": [
          {
            "description": "按媒体版本过滤",
            "in": "query",
            "name": "media_version_id",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionListResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "决策变更历史（追加式，不覆盖）",
        "tags": [
          "production_facts"
        ],
        "x-contract-source": "manual"
      }
    },
    "/api/production_facts/intents": {
      "get": {
        "operationId": "listProductionIntents",
        "parameters": [
          {
            "description": "按项目过滤",
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "按意图过滤",
            "in": "query",
            "name": "intent_id",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntentListResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "生成意图列表（冻结后的生成事实）",
        "tags": [
          "production_facts"
        ],
        "x-contract-source": "manual"
      },
      "post": {
        "operationId": "createProductionIntent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIntentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntentResponse"
                }
              }
            },
            "description": "成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "参数非法"
          }
        },
        "summary": "登记生成意图（冻结：此后不可改）",
        "tags": [
          "production_facts"
        ],
        "x-contract-source": "manual"
      }
    },
    "/api/production_facts/intents/{intent_id}": {
      "get": {
        "operationId": "getProductionIntent",
        "parameters": [
          {
            "description": "生成意图 id",
            "in": "path",
            "name": "intent_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntentResponse"
                }
              }
            },
            "description": "成功"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "意图不存在"
          }
        },
        "summary": "生成意图详情",
        "tags": [
          "production_facts"
        ],
        "x-contract-source": "manual"
      }
    },
    "/api/production_facts/intents/{intent_id}/derive": {
      "post": {
        "operationId": "deriveProductionIntent",
        "parameters": [
          {
            "description": "生成意图 id",
            "in": "path",
            "name": "intent_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntentResponse"
                }
              }
            },
            "description": "成功"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "意图不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "无实质变更，不予派生"
          }
        },
        "summary": "派生新意图（冻结意图不可原地改）",
        "tags": [
          "production_facts"
        ],
        "x-contract-source": "manual"
      }
    },
    "/api/production_facts/media": {
      "get": {
        "operationId": "listMediaVersions",
        "parameters": [
          {
            "description": "按意图过滤",
            "in": "query",
            "name": "intent_id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "是否带上 decision_state 五态",
            "in": "query",
            "name": "include_state",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaVersionListResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "候选媒体版本列表（含采用/批准状态）",
        "tags": [
          "production_facts"
        ],
        "x-contract-source": "manual"
      },
      "post": {
        "operationId": "registerMediaVersion",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterMediaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaVersionResponse"
                }
              }
            },
            "description": "成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "参数非法"
          }
        },
        "summary": "登记一次生成产出的候选",
        "tags": [
          "production_facts"
        ],
        "x-contract-source": "manual"
      }
    },
    "/api/production_facts/media/{media_version_id}": {
      "get": {
        "operationId": "getMediaVersion",
        "parameters": [
          {
            "description": "媒体版本 id",
            "in": "path",
            "name": "media_version_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaVersionResponse"
                }
              }
            },
            "description": "成功"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "媒体版本不存在"
          }
        },
        "summary": "媒体版本详情",
        "tags": [
          "production_facts"
        ],
        "x-contract-source": "manual"
      }
    },
    "/api/production_facts/media/{media_version_id}/approve": {
      "post": {
        "operationId": "approveMediaVersion",
        "parameters": [
          {
            "description": "媒体版本 id",
            "in": "path",
            "name": "media_version_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveMediaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            },
            "description": "成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "缺少授权者"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "授权者非人工主体（机器账号一律拒绝）"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "媒体版本不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "尚未采用 —— 采用不等于批准"
          }
        },
        "summary": "批准放行（人工决定；须先采用）",
        "tags": [
          "production_facts"
        ],
        "x-contract-source": "manual"
      }
    },
    "/api/production_facts/media/{media_version_id}/select": {
      "post": {
        "operationId": "selectMediaVersion",
        "parameters": [
          {
            "description": "媒体版本 id",
            "in": "path",
            "name": "media_version_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectMediaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            },
            "description": "成功"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "媒体版本不存在"
          }
        },
        "summary": "采用这版（创作决定；永不隐式升级为批准）",
        "tags": [
          "production_facts"
        ],
        "x-contract-source": "manual"
      }
    },
...
  },
  "servers": [
    {
      "description": "同源部署",
      "url": "/"
    }
  ],
...
}